added css

+++ fifa-test-working/models/Player.js
@@ -14,7 +14,7 @@ module.exports = function(sequelize, DataTypes) {
       }
     },
     strength: {
-      type: DataTypes.INTEGER
+      type: DataTypes.DECIMAL( 3 , 1 ),
     },
     priceTag: {
       type: DataTypes.STRING,
